# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  controllers.RegisterReq:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  models.User:
    properties:
      balance:
        type: number
      color:
        type: string
      email:
        type: string
      id:
        type: string
      initial:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        description: '"user", "admin", "manager", "teller", "operator"'
        type: string
      status:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: REST API for SmartBank Application
  title: SmartBank API
  version: "1.0"
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      description: Login and receive a JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
...
  /api/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: Registration Details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"